Conf: .load() Properties using a Reader, not an InputStream
The expectation that property files be encoded in ISO-8859-1 is a relic of the
past which has no justification in this day and age. UTF-8 is used almost
everywhere for text, with a notable exception: Windows operating systems, which
still insist on their own encodings, or as they call it, codepages (cp850,
cp1252).

The saving grace here is that the Java platform is good enough to accurately
detect the encoding used by a given platform that we can make use of it, using
Charset.defaultCharset(); also, Properties can .load() from a Reader since Java
6.

Make use of it.